+++ tailwind.config.cjs
@@ -1,20 +1,45 @@
-/** @type {import('tailwindcss').Config} */
 module.exports = {
     content: {
         relative: true,
         files: [
-            "./src/**/*.{js,ts,jsx,tsx}",
+            "./src/**/*.{js,ts,jsx,tsx}"
         ]
     },
     theme: {
         extend: {
             colors: {
-                primaryColor: "var(--primary-color)",
-                dangerColor: "var(--danger-color)",
-                textColor: "var(--text-color)",
-                borderColor: "var(--border-color)",
-                lightBackgroundColor: "var(--light-background-color)",
+                primary: {
+                    DEFAULT: "var(--color-primary)",
+                    foreground: "var(--color-primary)",
+                },
+                destructive: {
+                    DEFAULT: "var(--color-destructive)",
+                    foreground: "var(--color-destructive)",
+                },
+                background: "var(--color-background)",
+                border: "var(--color-border)"
             },
+            borderRadius: {
+                    DEFAULT: "calc(var(--radius)*1px)"
+            },
+            fontSize: {
+                generic: ['calc(var(--generic)*1px)', {lineHeight: '1.5'}],
+            },
+            width: {
+                icon: `calc(var(--font-generic)*2px)`
+            },
+            height: {
+                icon: "calc(var(--font-generic)*2px)"
+            },
+            padding: {
+                generic: "calc(var(--spacing-padding-y)*1px) calc(var(--spacing-padding-x)*1px) calc(var(--spacing-padding-y)*1px) calc(var(--spacing-padding-x)*1px)"//"calc(var(--spacing-padding-y)*1px) calc(var(--spacing-padding-x)*1px) calc(var(--spacing-padding-y)*1px) calc(var(--spacing-padding-x)*1px)",
+            },
+            margin: {
+                innerBlock: "calc(var(--spacing-block-inner-height)*1px)"
+            },
+            borderWidth: {
+                generic: "calc(var(--border-width)*1px)"
+            }
         },
     },
 };