+++ CHANGELOG.md
@@ -7,8 +7,19 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix phoneNumberField so phone number is not in the payload instead of always valuating it to null
+
 ### Added
+
 - Allow to list and remove trusted devices
+-
+## [1.32.2] - 2025-02-10
+
+### Fixed
+
+- Fix date field value format
 
 ## [1.32.1] - 2025-02-06
 
@@ -569,7 +580,9 @@ The eye icon is now correctly displayed in the Auth widget.
 
 First version of the SDK Web UI.
 
-[Unreleased]: https://github.com/ReachFive/identity-web-ui-sdk/compare/1.32.1...HEAD
+[Unreleased]: https://github.com/ReachFive/identity-web-ui-sdk/compare/1.32.2...HEAD
+
+[1.32.2]: https://github.com/ReachFive/identity-web-ui-sdk/compare/v1.32.1...v1.32.2
 
 [1.32.1]: https://github.com/ReachFive/identity-web-ui-sdk/compare/v1.32.0...v1.32.1